fix: depart Temp.tsx from newPage
